+++ clients/staking/src/rest-api/types/set-on-chain-yields-locked-product-redeem-option-response.ts
@@ -1,3 +1,5 @@
+/* tslint:disable */
+
 /**
  * Binance Staking REST API
  *
@@ -11,6 +13,16 @@
  * Do not edit the class manually.
  */
 
-export * from './eth-staking-api';
-export * from './on-chain-yields-api';
-export * from './sol-staking-api';
+/**
+ *
+ * @export
+ * @interface SetOnChainYieldsLockedProductRedeemOptionResponse
+ */
+export interface SetOnChainYieldsLockedProductRedeemOptionResponse {
+    /**
+     *
+     * @type {boolean}
+     * @memberof SetOnChainYieldsLockedProductRedeemOptionResponse
+     */
+    success?: boolean;
+}