+++ clients/staking/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 3.0.0 - 2025-07-01
+
+### Added (12)
+
+- `getOnChainYieldsLockedPersonalLeftQuota()` (`GET /sapi/v1/onchain-yields/locked/personalLeftQuota`)
+- `getOnChainYieldsLockedProductList()` (`GET /sapi/v1/onchain-yields/locked/list`)
+- `getOnChainYieldsLockedProductPosition()` (`GET /sapi/v1/onchain-yields/locked/position`)
+- `getOnChainYieldsLockedRedemptionRecord()` (`GET /sapi/v1/onchain-yields/locked/history/redemptionRecord`)
+- `getOnChainYieldsLockedRewardsHistory()` (`GET /sapi/v1/onchain-yields/locked/history/rewardsRecord`)
+- `getOnChainYieldsLockedSubscriptionPreview()` (`GET /sapi/v1/onchain-yields/locked/subscriptionPreview`)
+- `getOnChainYieldsLockedSubscriptionRecord()` (`GET /sapi/v1/onchain-yields/locked/history/subscriptionRecord`)
+- `onChainYieldsAccount()` (`GET /sapi/v1/onchain-yields/account`)
+- `redeemOnChainYieldsLockedProduct()` (`POST /sapi/v1/onchain-yields/locked/redeem`)
+- `setOnChainYieldsLockedAutoSubscribe()` (`POST /sapi/v1/onchain-yields/locked/setAutoSubscribe`)
+- `setOnChainYieldsLockedProductRedeemOption()` (`POST /sapi/v1/onchain-yields/locked/setRedeemOption`)
+- `subscribeOnChainYieldsLockedProduct()` (`POST /sapi/v1/onchain-yields/locked/subscribe`)
+
 ## 2.0.4 - 2025-06-19
 
 ### Changed (1)